requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-table==5.0.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==1.5.3
plotly==5.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
retrying==1.3.4
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2